Large number of implementation additions. Need to work on database storing jmodKeys
