+++ builtin/test/testmod.go
@@ -1,5 +1,3 @@
-
-
 package main
 
 import (
