+++ jablkoconfig.json
@@ -8,18 +8,15 @@
 				},
 				"inst1": {
 					"value": 32
-				},
-				"inst2": {
-					"value": 32
-				},
-				"inst3": {
-					"value": 32
-				},
-				"inst4": {
-					"value": 32
-				},
-				"inst5": {
-					"value": 32
+				}
+			}
+		},
+		"/home/coverstreet/Coding/Jablko_Home/My_Mods/Jarmuz-RGB-Light": {
+			"instances": {
+				"inst1": {
+					"lightIPs": [
+						"10.0.0.42"
+					]
 				}
 			}
 		}
